tests/requirements.txt
# 测试依赖包
# 核心测试框架
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0  # 并行测试
pytest-html>=3.1.0   # HTML报告
pytest-json-report>=1.5.0  # JSON报告

# AWS服务模拟
moto[all]>=4.2.0
boto3>=1.26.0
botocore>=1.29.0

# HTTP测试
requests>=2.28.0
responses>=0.22.0
httpx>=0.24.0

# 性能和内存监控
psutil>=5.9.0
memory-profiler>=0.60.0

# 数据处理和验证
pydantic>=1.10.0
jsonschema>=4.17.0

# 时间和日期处理
python-dateutil>=2.8.0
freezegun>=1.2.0  # 时间模拟

# 文件处理
python-pptx>=0.6.21  # PPT文件处理
Pillow>=9.0.0  # 图像处理

# 开发工具
black>=22.0.0  # 代码格式化
flake8>=5.0.0  # 代码检查
mypy>=1.0.0    # 类型检查

# 测试数据生成
factory-boy>=3.2.0
faker>=18.0.0

# 并发测试
pytest-concurrent>=0.1.0

# 测试覆盖率增强
coverage[toml]>=7.0.0

# API测试
tavern>=2.0.0  # API测试框架